src/churn_prediction_tbauctions/api/requirements.txt
scikit-learn>=1.4.1.post1
polars>=0.20.13
pandas>=2.2.1
pyarrow>=15.0.0
holidays>=0.44
fastapi>=0.110.0
uvicorn>=0.27.1
python-dotenv>=1.0.1
